requirement.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
distro==1.8.0
elevenlabs==0.2.26
executing==2.0.1
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
idna==3.5
ipython==8.18.0
jedi==0.19.1
matplotlib-inline==0.1.6
oauthlib==3.2.2
openai==1.3.5
parso==0.8.3
prompt-toolkit==3.0.41
protobuf==4.25.1
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.12
websockets==12.0
